agentic-rag-simple-demo/pyproject.toml
[project]
name = "agentic-rag-simple-demo"
version = "0.1.0"
description = "Agentic RAG - Agent decides when to retrieve from knowledge base"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.8.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
